Changes to the structure of the project
Changes to the file structure of the project and an introduction of the view for the user to input a file into the program. Checks are performed and error handling is implemented.
