+++ src/main/java/com/teamsix/employees/EmployeeManagerMain.java
@@ -1,0 +1,9 @@
+package com.teamsix.employees;
+
+import com.teamsix.employees.view.EmployeeView;
+
+public class EmployeeManagerMain {
+    public static void main(String[] args) {
+        EmployeeView.userFileInput();
+    }
+}
