+++ src/test/java/EmployeeReaderStepdefs.java
@@ -1,14 +1,12 @@
-import com.teamsix.employees.Employee;
-import com.teamsix.employees.EmployeeReader;
+import com.teamsix.employees.model.Employee;
+import com.teamsix.employees.model.EmployeeReader;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import org.junit.jupiter.api.Assertions;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 public class EmployeeReaderStepdefs {
     EmployeeReader reader;
